pyproject.toml
[tool.poetry]
name = "college-project-backend"
version = "0.1.0"
description = ""
authors = ["Narek Navoyan <narek.navoyan.05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0"
motor = "^3.3"
redis = "^5.0"
email-validator = "^2.1"
fastapi-mail = "^1.4"
python-multipart = "^0.0.9"
python-jose = {version = "^3.3", extras = ["cryptography"]}
passlib = {version = "^1.7", extras = ["bcrypt"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
